Distributor: split otlp related functions from push
